rename minnesota.h to mniw.h
